// frontend/src/api/generated/models/character-title-read.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Schema for reading a character title.
 * @export
 * @interface CharacterTitleRead
 */
export interface CharacterTitleRead {
    /**
     * 
     * @type {string}
     * @memberof CharacterTitleRead
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof CharacterTitleRead
     */
    'description': string;
    /**
     * 
     * @type {object}
     * @memberof CharacterTitleRead
     */
    'effects'?: object | null;
    /**
     * 
     * @type {boolean}
     * @memberof CharacterTitleRead
     */
    'is_equipped'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof CharacterTitleRead
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof CharacterTitleRead
     */
    'character_id': string;
    /**
     * 
     * @type {string}
     * @memberof CharacterTitleRead
     */
    'acquired_at': string;
    /**
     * 
     * @type {Date}
     * @memberof CharacterTitleRead
     */
    'created_at': Date;
    /**
     * 
     * @type {Date}
     * @memberof CharacterTitleRead
     */
    'updated_at': Date;
}